feature(CoWorker): can update
Closes #10
